api/pyproject.toml
[tool.poetry]
name = "ine-api-challenge"
version = "0.1.0"
description = "API Challenge"
authors = ["Martin Dimondo <martin.dimondo@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.black]
line-length = 79
target-version = ['py310']


[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.7"
djangorestframework = "^3.14.0"
psycopg2-binary = "^2.9.5"
django-oauth-toolkit = "^2.2.0"
requests = "^2.28.2"
responses = "^0.22.0"


[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
pylint = "^2.16.2"
pre-commit = "^3.0.4"
pytest = "^7.2.1"
pytest-django = "^4.5.2"
coverage = "^7.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
